+++ go/discovery_kit_api/CHANGELOG.md
@@ -1,8 +1,11 @@
 # Changelog
 
+## 1.5.1
+- Deprecate: DiscoveryDescription.restrictTo. Not needed anymore.
+
 ## 1.5.0
 - Target Ids must be unique across target type
--  Aligned Http Method constants
+- Aligned Http Method constants
 
 ## 1.4.2
 - Embed the openapi spec into the api package